+++ SPIRV/GlslangToSpv.h
@@ -35,8 +35,6 @@
 
 #pragma once
 
-#include "SpvTools.h"
-
 #include <string>
 #include <vector>
 
@@ -44,6 +42,18 @@
 
 namespace glslang {
 class TIntermediate;
+
+struct SpvOptions {
+    bool generateDebugInfo {false};
+    bool stripDebugInfo {false};
+    bool disableOptimizer {true};
+    bool optimizeSize {false};
+    bool disassemble {false};
+    bool validate {false};
+    bool emitNonSemanticShaderDebugInfo {false};
+    bool emitNonSemanticShaderDebugSource{ false };
+    bool compileOnly{false};
+};
 
 void GetSpirvVersion(std::string&);
 int GetSpirvGeneratorVersion();
